+++ keyboards/handwired/franky36/halconf.h
@@ -16,6 +16,6 @@
 
 #pragma once
 
-#define I2C_DRIVER I2CD0
-#define I2C1_SDA_PIN GP0
-#define I2C1_SCL_PIN GP1
+#define HAL_USE_I2C TRUE
+
+#include_next <halconf.h>
