+++ keyboards/handwired/franky36/keyboard.json
@@ -6,8 +6,6 @@
     "diode_direction": "COL2ROW",
     "features": {
         "bootmagic": true,
-        "command": false,
-        "console": false,
         "extrakey": true,
         "mousekey": true,
         "nkro": true,
@@ -18,7 +16,7 @@
         "rows": ["GP5", "GP4", "GP3", "GP2"]
     },
     "processor": "RP2040",
-    "url": "",
+    "url": "https://github.com/avdyushin/franky36",
     "usb": {
         "device_version": "0.2.0",
         "pid": "0x0001",
@@ -61,16 +59,12 @@
                 {"matrix": [2, 7], "x": 7, "y": 2},
                 {"matrix": [2, 8], "x": 8, "y": 2},
                 {"matrix": [2, 9], "x": 9, "y": 2},
-                {"matrix": [3, 0], "x": 0, "y": 3},
-                {"matrix": [3, 1], "x": 1, "y": 3},
                 {"matrix": [3, 2], "x": 2, "y": 3},
                 {"matrix": [3, 3], "x": 3, "y": 3},
                 {"matrix": [3, 4], "x": 4, "y": 3},
                 {"matrix": [3, 5], "x": 5, "y": 3},
                 {"matrix": [3, 6], "x": 6, "y": 3},
-                {"matrix": [3, 7], "x": 7, "y": 3},
-                {"matrix": [3, 8], "x": 8, "y": 3},
-                {"matrix": [3, 9], "x": 9, "y": 3}
+                {"matrix": [3, 7], "x": 7, "y": 3}
             ]
         }
     }
