--- keyboards/handwired/franky36/keymaps/default/rules.mk
@@ -1,1 +1,0 @@
-COMBO_ENABLE = yes
