--- keyboards/handwired/franky36/rules.mk
@@ -1,1 +1,0 @@
-OLED_DRIVER_ENABLE = yes
